+++ demos/VoronoiCell.js
@@ -1,15 +1,19 @@
 /**
- * A simple voronoi cell (part of a voronoi diagram), stored as an array of 
+ * @classdesc A simple voronoi cell (part of a voronoi diagram), stored as an array of 
  * adjacent triangles.
  *
- * @requires Triangle
+ * @requires Triangle, Polygon, Vertex
  *
  * @author   Ikaros Kappler
  * @date     2018-04-11
  * @modified 2018-05-04 Added the 'sharedVertex' param to the constructor. Extended open cells into 'infinity'.
  * @modified 2019-10-25 Fixed a serious bug in the toPathArray function; cell with only one vertex (extreme cases) returned invalid arrays which broke the rendering. 
  * @modified 2019-12-09 Removed an unnecesary if-condition from the _calculateOpenEdgePoint(...) helper function.
- * @version  1.0.2
+ * @modofied 2020-05-18 Added function VoronoiCell.toPolygon().
+ * @version  1.1.0
+ *
+ * @file VoronoiCell
+ * @public
  **/
 
 (function(context) {
@@ -43,6 +47,17 @@
 	return this.triangles.length < 3 || !this.triangles[0].isAdjacent(this.triangles[this.triangles.length-1]);	   
     };
 
+    /**
+     * Convert this Voronoi cell to a path polygon, consisting of all Voronoi cell corner points.
+     *
+     * Note that open Voronoi cell cannot properly converted to Polygons as they are considered
+     * infinite. 'Open' Voronoi edges will be cut off at their innermose vertices.
+     *
+     * @return {Polygon} 
+     **/
+    context.VoronoiCell.prototype.toPolygon = function() {
+	return new Polygon( this.toPathArray(), this.isOpen() );
+    };
 
     /**
      * Convert the voronoi cell path data to an SVG polygon data string.
@@ -109,13 +124,6 @@
 	var edgePoint = _findOuterEdgePoint( tri, neigh, sharedVertex );
 	var perpendicular = _perpendicularLinePoint( sharedVertex, edgePoint, center );
 	var openEdgePoint = null;
-	/* if( new Triangle(sharedVertex,center,edgePoint).determinant() <= 0 ) 
-	    openEdgePoint = new Vertex( perpendicular.x + (center.x-perpendicular.x)*1000,
-					perpendicular.y + (center.y-perpendicular.y)*1000 );
-	else
-	    openEdgePoint = new Vertex( perpendicular.x + (perpendicular.x-center.x)*1000,
-					perpendicular.y + (perpendicular.y-center.y)*1000 );
-	*/
 	// It is not necesary to make a difference on the determinant here
 	openEdgePoint = new Vertex( perpendicular.x + (center.x-perpendicular.x)*1000,
 				    perpendicular.y + (center.y-perpendicular.y)*1000 );